refactor(#255): 매칭 중 WaitingBox 컴포넌트 분리
